+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [feat:support reading configuration from application.yml or application.properties.](https://github.com/Tencent/spring-cloud-tencent/pull/262)
 - [fix:fix ClassNotFoundException while not importing openfeign when using circuit-breaker module.](https://github.com/Tencent/spring-cloud-tencent/pull/271)
 - [fix:solve ratelimit-callee-service UnknownHostException.](https://github.com/Tencent/spring-cloud-tencent/pull/292)
+- [Feature: Add config change listener feature support](https://github.com/Tencent/spring-cloud-tencent/pull/299)